Move utility methods out of extensions
Getting render to not need extension specific code but import only from util
